+++ src/types/index.ts
@@ -119,6 +119,7 @@ export type {
   Workout,
   WorkoutEx,
   WorkoutInput,
+  WorkoutConversionInput,
   WorkoutFormat,
   DuplicateWorkoutsDTO,
 } from './workout.js';